+++ tests/phpstan-bootstrap.php
@@ -12,4 +12,4 @@ defined( 'WP_CONTENT_DIR' ) || define( 'WP_CONTENT_DIR', '/tmp/wordpress/wp-cont
 defined( 'GTPERF_DIR' ) || define( 'GTPERF_DIR', dirname( __DIR__ ) );
 defined( 'GTPERF_FILE' ) || define( 'GTPERF_FILE', dirname( __DIR__ ) . '/gt-performance.php' );
 defined( 'GTPERF_BASENAME' ) || define( 'GTPERF_BASENAME', 'gt-performance/gt-performance.php' );
-defined( 'GTPERF_VERSION' ) || define( 'GTPERF_VERSION', '1.0.4' );
+defined( 'GTPERF_VERSION' ) || define( 'GTPERF_VERSION', '1.0.5' );